Changed price and amounts from Integer to Double so we can do the discounts correctly, the logic is now done too, but still not proved

+++ src/main/java/com/example/insurance/controller/CashRegisterController.java
@@ -18,7 +18,7 @@ public class CashRegisterController {
         return cashRegisterService.getCashRegisterByLastId();
     }
 
-    @PostMapping public String addMoneyToCashRegister(@RequestBody Integer money){
+    @PostMapping public String addMoneyToCashRegister(@RequestBody Double money){
         return cashRegisterService.addMoneyToCashRegister(money);
     }
